Better client and host key management
Use configured private key for the communication rather than the system keys
and publish the public key via API.

The command can be now provided with host public key to perform the
verification. And if not provided, it can extract the public key
from the host (saving into the action is not provided yet however).