Fixes #36615 - Resume dispatcher actor on exception
The previous implementation wrongly assumed that the :resume! part of
behaviour applied to the actor while in fact it applies to child actors
supervised by the actor with this behaviour.

This change spawns the dispatcher actor under a supervisor, which
resumes the dispatcher in case the dispatcher raises an exception.