Fixes #37500 - Improve cancellation on modern linux
Currently we spawn a somewhat deep hierarchy of processes.

For a simple "while true; sleep 1; done" we spawn something like this

bash(28973)─┬─sh(28992)───sudo(28995)───sudo(28999)───sh(29000)───sleep(29174)
            └─tee(28993)

As part of the preparation, we stored pid of the second-level
shell (28992 in this case). When the job is cancelled (be it by the user
or because the timeout to kill was reached), we send a SIGINT to all
processes which have the second-level shell as its parent - in this case
sudo(28995) and then rely on that process to either pass the signal to
its children and so on. Sadly this isn't completely reliable and the
leaf processes can stick around and just be re-parented under init.

Leaning into namespace support in linux, we can make this much more
reliable by isolating the entire subtree into a namespace and have all
processes inside it killed on cancellation.

bash(29913)─┬─sh(29932)───unshare(29934)───sudo(29935)───sudo(29937)───sh(29938)───sleep(29998)
            └─tee(29933)

With this patch, the subtree is a bit deeper. The added link there is
unshare, which is used to set up the namespace. Instead of storing a pid
of one of the shells, we store pid of unshare itself. On cancellation we
send SIGTERM to all its direct children. The expectation is that the
children (sudo in this case) will exit, which will in turn make unshare
exit and collapse the namespace and kill all the processes in it.