Fixes #38286 - Deploy script with tighter permissions (#124)
* Fixes #38286 - Deploy script with tighter permissions

* Refs #38286 - Ensure access for effective user

The files are deployed by the connection user, but they need to be
readable and executable by the by the effective user. Sadly, a non-root
user cannot change ownership of a file to another user so we have to
rely on ACLs to grant the permissions to the other user.

* Refs #38286 - Tighten permissions on the output file